+++ app/src/main/java/ly/count/android/demo/ActivityExampleCrashReporting.java
@@ -46,7 +46,10 @@ public class ActivityExampleCrashReporting extends Activity {
 
     public void onClickCrashReporting03(View v) {
         Countly.sharedInstance().addCrashBreadcrumb("Null pointer crash");
-        Countly.sharedInstance().crashTest(3);
+
+        Object[] o = null;
+        //noinspection InfiniteLoopStatement
+        while (true) { o = new Object[] { o }; }
     }
 
     public void onClickCrashReporting04(View v) {
@@ -66,7 +69,9 @@ public class ActivityExampleCrashReporting extends Activity {
     public void onClickCrashReporting07(View v) {
         Countly.sharedInstance().addCrashBreadcrumb("Custom crash log crash");
         Countly.sharedInstance().addCrashBreadcrumb("Adding some custom crash log");
-        Countly.sharedInstance().crashTest(2);
+
+        //noinspection UnusedAssignment,divzero
+        @SuppressWarnings("NumericOverflow") int test = 10/0;
     }
 
     public void onClickCrashReporting08(View v) {